renamed routes foler to appRoutes vsCode was using unusual folder icon when it was named routes, so renamed to be safe
